+++ test/simulation/go.mod
@@ -2,11 +2,12 @@ module github.com/ExchangeUnion/xud-simulation
 
 require (
 	github.com/Roasbeef/btcutil v0.0.0-20180406014609-dfb640c57141 // indirect
-	github.com/btcsuite/btcd v0.0.0-20190426011420-63f50db2f70a
+	github.com/btcsuite/btcd v0.0.0-20190614013741-962a206e94e9
 	github.com/frankban/quicktest v1.2.2 // indirect
 	github.com/go-errors/errors v1.0.1
 	github.com/golang/protobuf v1.3.1
 	github.com/google/go-cmp v0.3.0 // indirect
+	github.com/kkdai/bstream v1.0.0 // indirect
 	github.com/lightningnetwork/lnd v0.6.1-beta
 	github.com/ltcsuite/ltcd v0.0.0-20190507171044-fbadf835b5c0
 	github.com/ltcsuite/ltcutil v0.0.0-20190507133322-23cdfa9fcc3d
@@ -15,10 +16,12 @@ require (
 	github.com/roasbeef/btcd v0.0.0-20180418012700-a03db407e40d
 	github.com/roasbeef/btcutil v0.0.0-20180406014609-dfb640c57141
 	github.com/stretchr/testify v1.3.0
-	golang.org/x/crypto v0.0.0-20190513172903-22d7a77e9e5f // indirect
-	golang.org/x/net v0.0.0-20190514140710-3ec191127204
-	golang.org/x/sys v0.0.0-20190514135907-3a4b5fb9f71f // indirect
+	golang.org/x/crypto v0.0.0-20190618222545-ea8f1a30c443 // indirect
+	golang.org/x/net v0.0.0-20190620200207-3b0461eec859
+	golang.org/x/sys v0.0.0-20190620070143-6f217b454f45 // indirect
 	golang.org/x/text v0.3.2 // indirect
+	golang.org/x/tools v0.0.0-20190620191750-1fa568393b23 // indirect
+	google.golang.org/appengine v1.4.0
 	google.golang.org/genproto v0.0.0-20190513181449-d00d292a067c
 	google.golang.org/grpc v1.20.1
 	gopkg.in/macaroon.v2 v2.1.0
